preparing main

+++ libasm.h
@@ -5,11 +5,13 @@
 #include <unistd.h>
 #include <errno.h>
 #include <string.h>
+#include <stdlib.h>
+#include <fcntl.h>
 
-extern int ft_strlen(char *str);
+extern size_t ft_strlen(char *str);
 extern int ft_strcmp(char *s1, char *s2);
-extern int ft_write(int fd, void *buf, int byte);
-extern int ft_read(int fd, void *buf, int byte);
+extern size_t ft_write(int fd, void *buf, int byte);
+extern size_t ft_read(int fd, void *buf, int byte);
 extern char *ft_strcpy(char *dst, char *src);
 extern char *ft_strdup(char *str);
